refactor(smt): move RocksDbConfig to a shared file
for reuse across alternative backends
